Add support for deleting tutorials from students
When students are allocated a tutorial slot, deleting the tutorial
slot will not remove the allocation from students. As a result,
students are still shown to have the tutorial slot despite it not
existing in the application.

This causes confusion. For example, a user may try to delete the
tutorials, however they are unable to do so as the tutorials do not
exist. This can lead to inaccurate information.

Update the code to delete all tutorial allocations after a tutorial
has been deleted.